allow float amounts (e.g. 0.5 months)

+++ src/main/java/com/celements/invoice/builder/IInvoiceItem.java
@@ -18,8 +18,8 @@ public interface IInvoiceItem {
   public int getTotalPrice();
   public void setTotalPrice(int totalPrice);
   
-  public int getAmount();
-  public void setAmount(int amount);
+  public float getAmount();
+  public void setAmount(float amount);
   
   public String getCurrency();
   public void setCurrency(String currency);
